requirements-dev.txt
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.9
mypy>=1.8.0
pyinstaller>=6.0.0
pillow>=10.0.0